refactor(python): align retry module naming
